+++ R/class_aws_keras.R
@@ -14,7 +14,7 @@ aws_keras_new <- function(file = NULL, resources = NULL) {
       "tar_cloud",
       "tar_external",
       "tar_keras",
-      "tar_unexportable",
+      "tar_nonexportable",
       "tar_store"
     )
   )