+++ R/class_aws_fst.R
@@ -1,5 +1,5 @@
 #' @export
-store_new.aws_fst <- function(class, file = NULL, resources = NULL) {
+store_new.aws_fst <- function(format, file = NULL, resources = NULL) {
   aws_fst_new(file = file, resources = resources)
 }
 
@@ -20,5 +20,5 @@ aws_fst_new <- function(file = NULL, resources = NULL) {
 }
 
 #' @export
-store_assert_format_setting.aws_fst <- function(class) {
+store_assert_format_setting.aws_fst <- function(format) {
 }