+++ R/utils_oop.R
@@ -1,8 +1,0 @@
-as_class <- function(x) {
-  enclass(list(x), x)
-}
-
-enclass <- function(x, class) {
-  class(x) <- c(class, class(x))
-  x
-}